some changes

+++ src/main/java/com/example/demo/model/Tool.java
@@ -1,5 +1,7 @@
 package com.example.demo.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,7 +15,7 @@ public class Tool {
     private String toolCode;
     private String toolType;
     private String brand;
-    private double dailyCharge;
+    private BigDecimal dailyCharge;
     private boolean weekdayCharge;
     private boolean weekendCharge;
     private boolean holidayCharge;
